Fetch data once for main page
